fix(loaction): file name changed and provider IValidator and baseprovider implementeded
